Auto-run the acceptance test
Update the web service to return a 503 (unavailable)
when the api cannot be reached. A readiness probe was
also added to the k8s deployment to wait for a 2xx.

The acceptance test using the UI was updated to auto-run.
This means that a passing tilt ci check means the acceptance
test passes.